requirements.txt
Library We need for this project :
pandas
matplotlib
numpy
mysql-connector-python
spotipy
jupyter
